+++ src/GZCTF.Test/ConfigServiceTest.cs
@@ -7,15 +7,8 @@ using Xunit.Abstractions;
 
 namespace GZCTF.Test;
 
-public class ConfigServiceTest
+public class ConfigServiceTest(ITestOutputHelper output)
 {
-    readonly ITestOutputHelper output;
-
-    public ConfigServiceTest(ITestOutputHelper _output)
-    {
-        output = _output;
-    }
-
     [Fact]
     public void TestGetConfigs()
     {