+++ src/GZCTF/Extensions/TelemetryExtension.cs
@@ -17,7 +17,9 @@ public static class TelemetryExtension
 
         var otl = services.AddOpenTelemetry();
 
-        otl.ConfigureResource(resource => resource.AddService("GZCTF"));
+        otl.ConfigureResource(resource =>
+            resource.AddService("GZCTF",
+                serviceVersion: typeof(Program).Assembly.GetName().Version?.ToString(3)));
 
         otl.WithMetrics(metrics =>
         {
@@ -61,7 +63,8 @@ public static class TelemetryExtension
 
         if (config.OpenTelemetry.Enable)
         {
-            otl.UseOtlpExporter(config.OpenTelemetry.Protocol, new(config.OpenTelemetry.EndpointUri ?? "http://localhost:4317"));
+            otl.UseOtlpExporter(config.OpenTelemetry.Protocol,
+                new(config.OpenTelemetry.EndpointUri ?? "http://localhost:4317"));
         }
     }
 
@@ -70,10 +73,36 @@ public static class TelemetryExtension
         if (config is not { Enable: true, Prometheus.Enable: true })
             return;
 
+        if (config.Prometheus.Port is { } port)
+            app.UseOpenTelemetryPrometheusScrapingEndpoint(context =>
+            {
+                if (context.Connection.LocalPort == port
+                    && string.Equals(
+                        context.Request.Path.ToString().TrimEnd('/'),
+                        "/metrics",
+                        StringComparison.OrdinalIgnoreCase))
+                {
+                    // FIXME: workaround for prometheus
+                    context.Request.Headers.Accept = "application/openmetrics-text";
+                    return true;
+                }
 
-        if (config.Prometheus.Port is ushort port)
-            app.UseOpenTelemetryPrometheusScrapingEndpoint(context => context.Connection.LocalPort == port);
+                return false;
+            });
         else
-            app.UseOpenTelemetryPrometheusScrapingEndpoint();
+            app.UseOpenTelemetryPrometheusScrapingEndpoint(context =>
+            {
+                if (string.Equals(
+                    context.Request.Path.ToString().TrimEnd('/'),
+                    "/metrics",
+                    StringComparison.OrdinalIgnoreCase))
+                {
+                    // FIXME: workaround for prometheus
+                    context.Request.Headers.Accept = "application/openmetrics-text";
+                    return true;
+                }
+
+                return false;
+            });
     }
 }