+++ src/GZCTF/Services/Cache/CacheMaker.cs
@@ -32,7 +32,7 @@ public class CacheMaker(
     readonly Dictionary<string, ICacheRequestHandler> _cacheHandlers = new();
     CancellationTokenSource TokenSource { get; set; } = new();
 
-    public Task StartAsync(CancellationToken token)
+    public async Task StartAsync(CancellationToken token)
     {
         TokenSource = new CancellationTokenSource();
 
@@ -42,9 +42,7 @@ public class CacheMaker(
 
         #endregion
 
-        _ = Maker(TokenSource.Token);
-
-        return Task.CompletedTask;
+        await Task.Factory.StartNew(() => Maker(TokenSource.Token), token, TaskCreationOptions.LongRunning, TaskScheduler.Default);
     }
 
     public Task StopAsync(CancellationToken token)
@@ -69,7 +67,7 @@ public class CacheMaker(
         {
             await foreach (CacheRequest item in channelReader.ReadAllAsync(token))
             {
-                if (!_cacheHandlers.ContainsKey(item.Key))
+                if (!_cacheHandlers.TryGetValue(item.Key, out ICacheRequestHandler? handler))
                 {
                     logger.SystemLog(
                         Program.StaticLocalizer[nameof(Resources.Program.Cache_NoMatchingRequest), item.Key],
@@ -78,7 +76,6 @@ public class CacheMaker(
                     continue;
                 }
 
-                ICacheRequestHandler handler = _cacheHandlers[item.Key];
                 var key = handler.CacheKey(item);
 
                 if (key is null)