+++ src/GZCTF/Services/Interface/IMailSender.cs
@@ -1,4 +1,8 @@
-﻿namespace GZCTF.Services.Interface;
+﻿using GZCTF.Models.Internal;
+using Microsoft.Extensions.Localization;
+using Microsoft.Extensions.Options;
+
+namespace GZCTF.Services.Interface;
 
 public interface IMailSender
 {
@@ -23,7 +27,10 @@ public interface IMailSender
     /// <param name="userName">用户名</param>
     /// <param name="email">用户新注册的Email</param>
     /// <param name="confirmLink">确认链接</param>
-    public bool SendConfirmEmailUrl(string? userName, string? email, string? confirmLink);
+    /// <param name="localizer">本地化</param>
+    /// <param name="options">全局配置</param>
+    public bool SendConfirmEmailUrl(string? userName, string? email, string? confirmLink,
+        IStringLocalizer<Program> localizer, IOptionsSnapshot<GlobalConfig> options);
 
     /// <summary>
     /// 发送邮箱重置邮件
@@ -31,7 +38,10 @@ public interface IMailSender
     /// <param name="userName">用户名</param>
     /// <param name="email">用户的电子邮件</param>
     /// <param name="resetLink">重置链接</param>
-    public bool SendChangeEmailUrl(string? userName, string? email, string? resetLink);
+    /// <param name="localizer">本地化</param>
+    /// <param name="options">全局配置</param>
+    public bool SendChangeEmailUrl(string? userName, string? email, string? resetLink,
+        IStringLocalizer<Program> localizer, IOptionsSnapshot<GlobalConfig> options);
 
     /// <summary>
     /// 发送密码重置邮件
@@ -39,5 +49,8 @@ public interface IMailSender
     /// <param name="userName">用户名</param>
     /// <param name="email">用户的电子邮件</param>
     /// <param name="resetLink">重置链接</param>
-    public bool SendResetPasswordUrl(string? userName, string? email, string? resetLink);
+    /// <param name="localizer">本地化</param>
+    /// <param name="options">全局配置</param>
+    public bool SendResetPasswordUrl(string? userName, string? email, string? resetLink,
+        IStringLocalizer<Program> localizer, IOptionsSnapshot<GlobalConfig> options);
 }